Add files via upload
The above code helps to randomly select an anime from a list of given animes. It uses the choice function from the random module to randomly select an anime from the given list and print it. This is useful for when you want to randomly pick an anime to watch from a list of given options.